+++ pkg/client/tests/BUILD
@@ -13,6 +13,7 @@ go_test(
     srcs = [
         "fake_client_test.go",
         "listwatch_test.go",
+        "portfoward_test.go",
     ],
     library = ":go_default_library",
     tags = ["automanaged"],
@@ -23,13 +24,17 @@ go_test(
         "//pkg/api/v1:go_default_library",
         "//pkg/client/clientset_generated/internalclientset:go_default_library",
         "//pkg/client/clientset_generated/internalclientset/fake:go_default_library",
+        "//pkg/client/unversioned/remotecommand:go_default_library",
+        "//pkg/kubelet/server/portforward:go_default_library",
         "//vendor:k8s.io/apimachinery/pkg/apis/meta/v1",
         "//vendor:k8s.io/apimachinery/pkg/fields",
         "//vendor:k8s.io/apimachinery/pkg/runtime",
+        "//vendor:k8s.io/apimachinery/pkg/types",
         "//vendor:k8s.io/apimachinery/pkg/watch",
         "//vendor:k8s.io/client-go/pkg/api/install",
         "//vendor:k8s.io/client-go/rest",
         "//vendor:k8s.io/client-go/tools/cache",
+        "//vendor:k8s.io/client-go/tools/portforward",
         "//vendor:k8s.io/client-go/util/testing",
     ],
 )
